Validate destination URL on availability when saving a new redirect. Re #424

+++ classes/class.JTL-Shop.Redirect.php
@@ -125,7 +125,9 @@ class Redirect
             $cDestination = '/' . $cDestination;
         }
 
-        if (strlen($cSource) > 1 && strlen($cDestination) > 1 && $cSource != $cDestination || $bForce) {
+        if (self::checkAvailability(Shop::getURL() . $cDestination) &&
+            strlen($cSource) > 1 && strlen($cDestination) > 1 && $cSource != $cDestination || $bForce)
+        {
             if ($this->isDeadlock($cSource, $cDestination)) {
                 Shop::DB()->delete('tredirect', ['cToUrl', 'cFromUrl'], [$cSource, $cDestination]);
             }
